requirements.txt
aiosqlite
argon2-cffi
dummy-text-generator
faker
fastapi[standard]
fastapi_auth_middleware
pyjwt
sqlactive
